ATHENA_IDR_Classifier/requirements.txt
torch>=2.0.0
numpy>=1.21.0
biopython>=1.79
PyYAML>=6.0
